Change headers on requests for auth middleware

+++ src/redux/track/actions.js
@@ -7,28 +7,22 @@ export const trackRequest = (tracks) => ({
   payload: tracks,
 })
 
-export function getTracks(liked) {
+export function getTracks(liked, userId) {
   return async function getTracksThunk(dispatch) {
     const token = await auth.getCurrentUserToken()
     if (!token) return
-    const response = await api.getTracks(
-      { Authorization: `Bearer ${token}` },
-      liked
-    )
+    const response = await api.getTracks({ _id: userId }, liked)
 
     if (response.errorMessage) return
     dispatch(trackRequest(response.data.data))
   }
 }
 
-export function deleteTracks(id) {
+export function deleteTracks(id, userId) {
   return async function deleteTracksThunk(dispatch) {
     const token = await auth.getCurrentUserToken()
     if (!token) return
-    const response = await api.deleteTrack(
-      { Authorization: `Bearer ${token}` },
-      id
-    )
+    const response = await api.deleteTrack({ _id: userId }, id)
     if (response.errorMessage) return
     dispatch(renderTracks(response.data.data))
   }
